backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9
pgvector>=0.2.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
httpx>=0.27.0
numpy>=1.26.4
scipy>=1.13.0
Pillow>=10.3.0
opencv-python-headless>=4.9.0.80
deepface>=0.0.91
tf-keras>=2.16.0
speechbrain>=1.0.0
torch>=2.2.2
torchaudio>=2.2.2
soundfile>=0.12.1
librosa>=0.10.2
aiofiles>=23.2.1
